+++ docs/changelog.md
@@ -10,7 +10,11 @@ and this project adheres to [Effort-based Versioning](https://jacobtomlinson.dev
 
 ## [Unreleased][]
 
-(no summary yet)
+(no changes yet)
+
+## [3.0.2][] - 2025-06-14 {: v3.0.2 }
+
+Small improvements and a wrapper for git commands that depend on the commit id.
 
 ### Added
 
@@ -414,6 +418,7 @@ This is the first release of StepUp RepRep that is compatible with StepUp Core 2
 Initial release
 
 [Unreleased]: https://github.com/reproducible-reporting/stepup-reprep
+[3.0.2]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v3.0.2
 [3.0.1]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v3.0.1
 [3.0.0]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v3.0.0
 [2.3.6]: https://github.com/reproducible-reporting/stepup-reprep/releases/tag/v2.3.6